move path implementation helpers into separate subdirectory
