+++ src/commands/modifypointscommand.cpp
@@ -2,8 +2,8 @@
 #include "common.h"
 #include "scene/scene.h"
 #include "objects/path.h"
-#include "objects/pathpoint.h"
-#include "objects/segment.h"
+#include "objects/path/pathpoint.h"
+#include "objects/path/segment.h"
 
 namespace omm
 {
