+++ src/commands/subdividepathcommand.cpp
@@ -1,7 +1,7 @@
 #include "commands/subdividepathcommand.h"
 #include "objects/path.h"
-#include "objects/pathpoint.h"
-#include "objects/segment.h"
+#include "objects/path/pathpoint.h"
+#include "objects/path/segment.h"
 
 namespace
 {
