+++ src/mainwindow/pathactions.cpp
@@ -13,8 +13,8 @@
 #include "properties/optionproperty.h"
 #include "properties/referenceproperty.h"
 #include "objects/path.h"
-#include "objects/pathpoint.h"
-#include "objects/segment.h"
+#include "objects/path/pathpoint.h"
+#include "objects/path/segment.h"
 #include "objects/convertedobject.h"
 #include "scene/history/historymodel.h"
 #include "scene/history/macro.h"
