+++ src/objects/cloner.cpp
@@ -4,7 +4,7 @@
 
 #include "objects/convertedobject.h"
 #include "objects/empty.h"
-#include "objects/enhancedpathvector.h"
+#include "objects/path/enhancedpathvector.h"
 #include "properties/boolproperty.h"
 #include "properties/floatproperty.h"
 #include "properties/floatvectorproperty.h"
