+++ src/objects/empty.cpp
@@ -1,5 +1,5 @@
 #include "objects/empty.h"
-#include "objects/enhancedpathvector.h"
+#include "objects/path/enhancedpathvector.h"
 #include "properties/boolproperty.h"
 
 namespace omm
