+++ src/objects/lineobject.cpp
@@ -1,8 +1,8 @@
 #include "objects/lineobject.h"
 #include "properties/boolproperty.h"
 #include "properties/floatproperty.h"
-#include "objects/segment.h"
-#include "objects/enhancedpathvector.h"
+#include "objects/path/segment.h"
+#include "objects/path/enhancedpathvector.h"
 #include <QObject>
 
 namespace omm
