+++ src/objects/object.cpp
@@ -3,9 +3,9 @@
 #include "common.h"
 #include "logging.h"
 #include "objects/convertedobject.h"
-#include "objects/enhancedpathvector.h"
 #include "objects/path.h"
-#include "objects/segment.h"
+#include "objects/path/enhancedpathvector.h"
+#include "objects/path/segment.h"
 #include "properties/boolproperty.h"
 #include "properties/floatproperty.h"
 #include "properties/floatvectorproperty.h"
