+++ src/objects/path/CMakeLists.txt
@@ -1,0 +1,8 @@
+target_sources(libommpfritt PRIVATE
+  enhancedpathvector.cpp
+  enhancedpathvector.h
+  pathpoint.cpp
+  pathpoint.h
+  segment.cpp
+  segment.h
+)
