+++ src/objects/path/enhancedpathvector.h
