+++ src/objects/path/pathpoint.cpp
@@ -1,5 +1,5 @@
-#include "objects/pathpoint.h"
-#include "objects/segment.h"
+#include "objects/path/pathpoint.h"
+#include "objects/path/segment.h"
 #include "objects/path.h"
 #include "scene/disjointpathpointsetforest.h"
 #include "scene/scene.h"
