+++ src/objects/path/segment.h
