+++ src/objects/proceduralpath.cpp
@@ -1,9 +1,9 @@
 #include "objects/proceduralpath.h"
 #include "external/pybind11/stl.h"
 #include "objects/path.h"
-#include "objects/pathpoint.h"
-#include "objects/segment.h"
-#include "objects/enhancedpathvector.h"
+#include "objects/path/pathpoint.h"
+#include "objects/path/segment.h"
+#include "objects/path/enhancedpathvector.h"
 #include "properties/boolproperty.h"
 #include "properties/integerproperty.h"
 #include "properties/stringproperty.h"
