+++ src/objects/text.cpp
@@ -1,7 +1,7 @@
 #include "objects/text.h"
 #include "mainwindow/viewport/viewport.h"
 #include "objects/path.h"
-#include "objects/enhancedpathvector.h"
+#include "objects/path/enhancedpathvector.h"
 #include "properties/floatproperty.h"
 #include "properties/stringproperty.h"
 #include "renderers/painter.h"
