+++ src/tools/pathtool.cpp
@@ -4,8 +4,8 @@
 #include "commands/joinpointscommand.h"
 #include "main/application.h"
 #include "objects/path.h"
-#include "objects/pathpoint.h"
-#include "objects/segment.h"
+#include "objects/path/pathpoint.h"
+#include "objects/path/segment.h"
 #include "scene/scene.h"
 #include "scene/history/historymodel.h"
 #include "scene/history/macro.h"
